+++ romsearch/modules/__init__.py
@@ -4,6 +4,7 @@ from .gamefinder import GameFinder
 from .rahasher import RAHasher
 from .romchooser import ROMChooser
 from .romcleaner import ROMCleaner
+from .romcompressor import ROMCompressor
 from .romdownloader import ROMDownloader
 from .rommover import ROMMover
 from .romparser import ROMParser
@@ -17,6 +18,7 @@ __all__ = [
     "RAHasher",
     "ROMChooser",
     "ROMCleaner",
+    "ROMCompressor",
     "ROMDownloader",
     "ROMMover",
     "ROMParser",